Plugin: Move API.h/hpp into reframework/ subdirectory
