+++ example_plugin/Plugin.cpp
@@ -1,6 +1,6 @@
 #include <Windows.h>
 
-#include "API.hpp"
+#include <reframework/API.hpp>
 
 lua_State* g_lua = nullptr;
