+++ src/mods/PluginLoader.cpp
@@ -3,8 +3,8 @@
 #include <imgui.h>
 #include <spdlog/spdlog.h>
 
-#include "API.hpp"
 #include "REFramework.hpp"
+#include "reframework/API.hpp"
 
 #include "PluginLoader.hpp"
